+++ tests/reduce_ops_test.py
@@ -2,17 +2,13 @@
 # Copyright (c) 2022 Graphcore Ltd. All rights reserved.
 import torch
 import pytest
-import poptorch
 import helpers
+import poptorch
 
 
 # Reduce Ops Harness
 # Checks that the IPU reduce ops match the CPU version.
-def reduce_harness(trace_model,
-                   func,
-                   input,
-                   dim=None,
-                   expected_dtype=torch.bool):
+def reduce_harness(trace_model, func, input, dim=None):
     # dim must be passed this way to avoid named tensor errors
     kwargs = {"dim": dim} if dim else {}
 
@@ -21,7 +17,7 @@ def reduce_harness(trace_model,
 
     class Model(torch.nn.Module):
         def __init__(self):
-            super(Model, self).__init__()
+            super().__init__()
             self.reduce_op = reduce_op
 
         def forward(self, x):
@@ -36,10 +32,7 @@ def reduce_harness(trace_model,
     opts.Jit.traceModel(trace_model)
     pop_model = poptorch.inferenceModel(model, opts)
     pop_out = pop_model(input)
-    assert pop_out.dtype == expected_dtype
-
     native_out = model(input)
-    assert native_out.size() == pop_out.size()
 
     if torch.is_floating_point(native_out):
         helpers.assert_allclose(expected=native_out, actual=pop_out)
@@ -61,4 +54,12 @@ def test_any_all(trace_model, func, dim):
 @pytest.mark.parametrize("trace_model", [True, False])
 def test_sum_mean(trace_model, func, dim):
     input = torch.rand(32, 128)
-    reduce_harness(trace_model, func, input, dim, expected_dtype=torch.float32)
+    reduce_harness(trace_model, func, input, dim)
+
+
+@pytest.mark.parametrize("dim", (None, 0, -1, [1, 2]))
+@pytest.mark.parametrize("trace_model", [True, False])
+def test_count_nonzero(trace_model, dim):
+    torch.manual_seed(42)
+    input = torch.randint(10, (2, 3, 4, 5))
+    reduce_harness(trace_model, torch.count_nonzero, input, dim)